Move PIXI into React component
