+++ package.json
@@ -17,18 +17,40 @@
   "bugs": {
     "url": "https://github.com/chrismdp/tinyking/issues"
   },
+  "babel": {
+    "presets": [
+      [
+        "@babel/preset-env",
+        {
+          "targets": {
+            "browsers": [
+              "defaults"
+            ]
+          }
+        }
+      ],
+      "@babel/preset-react"
+    ]
+  },
   "homepage": "https://github.com/chrismdp/tinyking#readme",
   "devDependencies": {
     "@babel/core": "^7.10.4",
     "@babel/preset-env": "^7.10.4",
+    "@babel/preset-react": "^7.10.4",
     "babel-loader": "^8.1.0",
     "clean-webpack-plugin": "^3.0.0",
     "core-js": "^3.6.5",
     "css-loader": "^3.6.0",
     "eslint": "^7.4.0",
+    "eslint-import-resolver-webpack": "^0.12.2",
     "eslint-loader": "^4.0.2",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-react": "^7.20.3",
     "file-loader": "^6.0.0",
     "html-webpack-plugin": "^4.3.0",
+    "html-webpack-root-plugin": "git+https://github.com/jeremy-hunter/html-webpack-root-plugin.git#patch-1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "sass": "^1.26.10",
     "sass-loader": "^9.0.2",
     "style-loader": "^1.2.1",
