+++ src/style.scss
@@ -2,3 +2,10 @@ body {
   padding: 0;
   margin: 0;
 }
+
+canvas {
+  position: absolute;
+  top: 0px;
+  left: 0px;
+  z-index: -1;
+}
